backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djangorestframework-stubs
django-cors-headers
django-cognito-jwt
boto3
boto3-stubs
mysqlclient
stripe
celery[redis]
django-celery-beat
daphne
channels
python-dotenv